Moved theme management for popups to mixin as it was duplicated code.
