+++ module/ui/popups/feature-popup.mjs
@@ -1,17 +1,11 @@
 import { EventidePopupHelpers } from "../components/_module.mjs";
-import {
-  initThemeManager,
-  THEME_PRESETS,
-  applyThemeImmediate,
-  cleanupThemeManager,
-} from "../../helpers/_module.mjs";
-import { Logger } from "../../services/_module.mjs";
+import { ThemeManagedPopupMixin } from "../mixins/_module.mjs";
 
 /**
  * Application for displaying feature information in a popup.
- * @extends {EventidePopupHelpers}
+ * @extends {ThemeManagedPopupMixin(EventidePopupHelpers)}
  */
-export class FeaturePopup extends EventidePopupHelpers {
+export class FeaturePopup extends ThemeManagedPopupMixin(EventidePopupHelpers) {
   /** @override */
   static PARTS = {
     featurePopup: {
@@ -47,72 +41,6 @@ export class FeaturePopup extends EventidePopupHelpers {
   constructor({ item }) {
     super({ item });
     this.type = "feature";
-  }
-
-  /**
-   * Handle rendering of the feature popup application
-   * @param {ApplicationRenderContext} context      Prepared context data
-   * @param {RenderOptions} options                 Provided render options
-   * @protected
-   */
-  _onRender(_context, _options) {
-    super._onRender(_context, _options);
-
-    // Re-apply themes on re-render (but don't reinitialize)
-    if (this.themeManager) {
-      this.themeManager.applyThemes();
-    }
-  }
-
-  /**
-   * Handle the first render of the feature popup application
-   * @override
-   * @protected
-   */
-  async _onFirstRender() {
-    super._onFirstRender();
-
-    // Apply theme immediately to prevent flashing
-    applyThemeImmediate(this.element);
-
-    // Initialize theme management only on first render (non-blocking like actor/item sheets)
-    if (!this.themeManager) {
-      initThemeManager(this, THEME_PRESETS.CREATOR_APPLICATION)
-        .then((manager) => {
-          this.themeManager = manager;
-          Logger.debug(
-            "Theme management initialized asynchronously for feature popup",
-            {
-              hasThemeManager: !!this.themeManager,
-              sheetId: this.id,
-            },
-            "THEME",
-          );
-        })
-        .catch((error) => {
-          Logger.error(
-            "Failed to initialize theme manager for feature popup",
-            error,
-            "THEME",
-          );
-        });
-    }
-  }
-
-  /**
-   * Clean up resources before closing the application
-   * @param {Object} options - The options for closing
-   * @returns {Promise<void>}
-   * @override
-   */
-  async _preClose(options) {
-    // Clean up theme management for this specific instance
-    if (this.themeManager) {
-      cleanupThemeManager(this);
-      this.themeManager = null;
-    }
-
-    await super._preClose(options);
   }
 
   /**
